Rename filter module to filtering to avoid similarity to built-in "filter"
