+++ coursera/filtering.py
