+++ coursera/test/test_filter.py
@@ -1,4 +1,4 @@
-from coursera.filter import skip_format_url
+from coursera.filtering import skip_format_url
 
 
 def test_filter():
